Creates reducer for LIST ARTICLES action
Here, I am going with Option 2 from http://nikolay.rocks/2017-06-18-better-redux-loading,
which isn't 100% recommended but I'll go with it for now.

I do think the issue of "what if `loading: true` and `data: [ ...items]`?
How do you handle that "split" state?

I may look to create a new file in my redux directory: initialStates.js
and export all of those to be used in my reducers as well as my components.

It would be nice to not have to re-instantiate a new object every time
I want to reduce my action to a new state...

I'll look to implementing that at a later date.